requirements.txt
blinker==1.8.2
click==8.1.7
entrypoints==0.3
flake8==3.7.9
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
importlib_metadata==7.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
ldap3==2.9.1
MarkupSafe==2.1.5
mccabe==0.6.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyYAML==6.0.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
WTForms==3.1.2
zipp==3.19.2
